+++ Data/DataAccess/Queries/GetProductByIdQuery.cs
@@ -1,12 +1,12 @@
-﻿using MediatR;
-using Review.API.Common;
-using Review.API.Model;
-using Review.API.Repository.Interface;
+﻿using Data.Common;
+using Data.Model;
+using Data.Repository.Interface;
+using MediatR;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Review.API.DataAccess.Queries
+namespace Data.DataAccess.Queries
 {
     public class GetProductByIdQuery : IRequest<RequestResult<ProductModel>>
     {
