+++ Data/Repository/Implementation/ProductRepository.cs
@@ -1,11 +1,11 @@
-﻿using Review.API.DatabaseConfigurations;
-using Review.API.Model;
-using Review.API.Repository.Interface;
+﻿using Data.Model;
+using Data.Repository.Interface;
+using Review.API.DatabaseConfigurations;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Review.API.Repository.Class
+namespace Data.Repository.Implementation
 {
     public class ProductRepository : IProductRepository
     {
