+++ Data/Repository/Interface/IReviewRepository.cs
@@ -1,8 +1,8 @@
-﻿using Review.API.Model;
+﻿using Data.Model;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Review.API.Repository
+namespace Data.Repository.Interface
 {
     public interface IReviewRepository
     {
