+++ Review.API/Controllers/ReviewController.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Threading.Tasks;
+using Data.DataAccess.Command;
+using Data.DataAccess.Queries;
+using Data.Model;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
-using Review.API.DataAccess.Command;
-using Review.API.DataAccess.Queries;
 
 namespace Review.API.Controllers
 {
