+++ ReviewProduct.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.28729.10
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Review.API", "Review.API\Review.API.csproj", "{B601931E-FE2B-44D1-879D-C69CAB7217CF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data", "Data\Data.csproj", "{B5FDA05C-AFE0-4ED7-9B74-6B225944B861}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{B601931E-FE2B-44D1-879D-C69CAB7217CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B601931E-FE2B-44D1-879D-C69CAB7217CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B601931E-FE2B-44D1-879D-C69CAB7217CF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5FDA05C-AFE0-4ED7-9B74-6B225944B861}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5FDA05C-AFE0-4ED7-9B74-6B225944B861}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5FDA05C-AFE0-4ED7-9B74-6B225944B861}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5FDA05C-AFE0-4ED7-9B74-6B225944B861}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
